made-delete-employee-method
Josef and i made this

+++ src/Domain/SocialWorker.java
@@ -11,7 +11,7 @@ package Domain;
  */
 public class SocialWorker extends CaseEmployee{
     
-    public SocialWorker(int cpr, String name, char gender, String birthDate, String address, int phoneNumber, String mail, int id, String userName, String password) {
+    public SocialWorker(long cpr, String name, char gender, String birthDate, String address, int phoneNumber, String mail, int id, String userName, String password) {
         super(cpr, name, gender, birthDate, address, phoneNumber, mail, id, userName, password);
     }
